+++ src/10_styles/00_base-styles/_zindex.scss
@@ -1,0 +1,11 @@
+/*
+ * Header Styles
+ *
+ * Sass Maps: https://sass-lang.com/documentation/values/maps
+ */
+
+$zindex: (
+    skiplink: 10,
+    header: 5,
+    footer: 5,
+);
