+++ src/10_styles/01_atom-styles/_a-ex-link.scss
@@ -1,4 +1,4 @@
-.m-ex-link {
+.a-ex-link {
     color: $link-color;
 
     svg {
